requirements.txt
nltk==3.8.1
PyMuPDF==1.21.1
camelot-py==0.11.0